sessions_controller and omniauth_controller changes
