+++ app/controllers/api/sessions_controller.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
-
 module Api
-    class SessionController < ApiController
+    class SessionsController < ApiController
       def create
   	     #@user = User.find_or_create_from_auth_hash(env["omniauth.auth"])
   	     #session[:user_id] = @user.id
@@ -10,11 +9,9 @@ module Api
          session[:user_id] = user.id
          redirect_to root_path
       end
-
       def destroy
         session[:user_id] = nil
         redirect_to root_path
       end
     end
-
 end
