--- app/controllers/users/omniauth_callbacks_controller.rb
@@ -1,27 +1,0 @@
-module Users
-  class OmniauthCallbacksController < Devise::OmniauthCallbacksController
-    # rubocop:disable AbcSize
-    def google_oauth2
-      user = User.from_omniauth(request.env['omniauth.auth'])
-      if user.persisted?
-        flash[:notice] = I18n.t 'devise.omniauth_callbacks.success', kind: 'Google'
-        sign_in_and_redirect user, event: :authentication
-      else
-        session['devise.google_data'] = request.env['omniauth.auth'].except(:extra)
-        redirect_to new_user_registration_url, alert: user.errors.full_messages.join("\n")
-      end
-    end
-  end
-  def passthru
-    render :file => "#{Rails.root}/public/404.html", :status => 404, :layout => false
-  end
-
-  def after_sign_in_path_for(resource)
-    if resource.email_verified?
-      super resource
-    else
-      finish_signup_path(resource)
-    end
-  end
-
-end
